backend/requirements.txt
fastapi[all]>=0.109.2
langgraph>=0.0.66
langchain-groq>=0.1.5
langchain>=0.2.6
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
redis>=5.0.4
python-telegram-bot>=21.1.1
duckduckgo-search>=6.1.7
httpx>=0.27.0
python-dotenv>=1.0.1
